+++ src/javascript/customization.js
@@ -73,6 +73,14 @@ const loadTheme = number => {
   for (i = 0; i < editors.length; i++) {
     if (editors[i].editor != undefined) editors[i].editor.setOption("theme", themes[number]["Highlight"]); //Update highlither after applying a new theme
   }
+  if(current_screen!=undefined){
+    if(current_screen.terminal!=undefined){
+      current_screen.terminal.xterm.setOption('theme', { 
+        background:themeObject.Colors["editor-background-color"],
+        foreground:themeObject.Colors["white-black"]
+      });
+    }
+  }
   current_config.theme = themes[number];
   saveConfig(); //Save the current configuration
 }
@@ -95,6 +103,16 @@ const setThemeByName = name => {
       }
       for (i = 0; i < editors.length; i++) {
         editors[i].editor.setOption("theme", themes[i]["Highlight"]); //Update highlither after applying a new theme
+      }
+      if(current_screen!=undefined){
+        if(current_screen.terminal!=undefined){
+          current_screen.terminal.xterm.setOption("theme",
+            {
+              background:themeObject.Colors["editor-background-color"],
+              foreground:themeObject.Colors["white-black"]
+            }
+          )
+        }
       }
       return;
     }